Condense e2e chat tests into main test suite for better performance
Change-type: patch
